# internal/webapi/openapi.yaml
openapi: 3.0.3
info:
  title: debuginfod-go HTTP API
  description: |
    Совместимый с elfutils debuginfod HTTP API.
    Целевые ОС: Astra Linux, Ubuntu, RedOS, CentOS.
  version: 1.0.0
  license:
    name: MIT
servers:
  - url: http://localhost:8002
    description: Локальный сервер
paths:
...
  /metadata:
    get:
      summary: Поиск артефактов
      parameters:
        - name: key
          in: query
          required: true
          schema:
            type: string
            enum: [glob, file, buildid]
        - name: value
          in: query
          required: true
          schema:
            type: string
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
      responses:
        "200":
          description: Результаты поиска
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetadataResponse"
        "400":
          description: Неверные параметры
...
  /admin/rescan:
    post:
      summary: Внеочередной rescan
      security:
        - AdminToken: []
      responses:
        "200":
          description: accepted
        "401":
          description: unauthorized
        "403":
          description: admin API disabled
        "409":
          description: scan disabled
  /zabbix:
    get:
      summary: Метрики для Zabbix HTTP agent
      security:
        - ZabbixToken: []
      responses:
        "200":
          description: JSON-метрики
...
components:
...
  schemas:
    ArtifactRecord:
      type: object
      properties:
        buildid:
          type: string
        type:
          type: string
          enum: [executable, debuginfo]
        file:
          type: string
        archive:
          type: string
        buildid_kind:
          type: string
        raw_buildid:
          type: string
    MetadataResponse:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/ArtifactRecord"
        complete:
          type: boolean
        next_offset:
          type: integer
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    ZabbixToken:
      type: apiKey
      in: header
      name: X-Zabbix-Token
    AdminToken:
      type: apiKey
      in: header
      name: X-Admin-Token